requirements.txt
Django==1.3.1
django-taggit==0.9.3
